+++ Changes.md
@@ -68,8 +68,10 @@ Breaking Changes
 - Handle : `LinearDrag::position` and `PlanarDrag::position` are no longer `const` methods. `RotateHandle`, `ScaleHandle` and `TranslateHandle` value provider methods loose `const`-ness accordingly.
 - ImagePlug : Removed `image()` and `imageHash()` methods. These are now available in the ImageAlgo namespace.
 - ImageNode : Added virtual methods.
+- Light :
+  - Changed `Light::computeLight` return type to const (#3407).
+  - Added plug and accessors for `visualiserShaded` (#3407).
 - StandardLightVisualiser : Removed protected members `faceCameraVertexSource` and `environmentSphere` (#3407).
-- Light : Added plugs and accessors for `visualiserShaded` (#3407).
 
 0.55.x.x (relative to 0.55.1.0)
 ========